separated two files, changed a not elegant object copy
